requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.17
Flask==3.1.0
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-datastore==2.3.0
google-cloud-storage==1.42.3
google-cloud-tasks==2.7.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.12.7
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
libcst==1.6.0
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.25.0
protobuf==3.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pytz==2021.1
PyYAML==6.0.2
redis==4.1.0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
